+++ src/collective/newsflash/setuphandlers.py
@@ -1,16 +1,27 @@
 # -*- coding: utf-8 -*-
+from Products.CMFPlone import interfaces as Plone
+from Products.CMFQuickInstallerTool import interfaces as QuickInstaller
+from zope.interface import implements
 
-from five import grok
 
-from Products.CMFPlone.interfaces import INonInstallable
+class HiddenProfiles(object):
 
-
-class HiddenProfiles(grok.GlobalUtility):
-
-    grok.implements(INonInstallable)
-    grok.provides(INonInstallable)
-    grok.name('collective.newsflash')
+    implements(Plone.INonInstallable)
 
     def getNonInstallableProfiles(self):
-        profiles = ['collective.newsflash:uninstall', ]
-        return profiles
+        """Do not show on Plone's list of installable profiles."""
+        return [
+            u'collective.newsflash:uninstall',
+            u'collective.newsflash.upgrades.v1001:default',
+        ]
+
+
+class HiddenProducts(object):
+
+    implements(QuickInstaller.INonInstallable)
+
+    def getNonInstallableProducts(self):
+        """Do not show on QuickInstaller's list of installable products."""
+        return [
+            u'collective.newsflash.upgrades.v1001'
+        ]